+++ src/mtool/api/UnitTests/POS_Operations/test_pos.py
@@ -104,10 +104,9 @@ def test_is_ibofos_runningi_failure(mock_get_current_user, mock_get_rebuilding_s
 """
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
 @mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
 @mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
-def test_start_ibofos(mock_get_current_user,mock_get_ibof_os_status,mock_is_ibofos_running, **kwargs):
+def test_start_ibofos(mock_get_current_user,mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
                        json={"result": {"status": {"description": "SUCCESS", "code":0}}},
@@ -165,10 +164,9 @@ def test_get_pos_info_err(mock_get_current_user, **kwargs):
     assert response.status_code == 500
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
 @mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
 @mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=True, autospec=True)
-def test_start_ibofos_already_running_failure(mock_get_current_user, mock_get_ibof_os_status, mock_is_ibofos_running, **kwargs):
+def test_start_ibofos_already_running_failure(mock_get_current_user, mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
                        json={"result": {"status": {"description": "", "code":0}}}, status_code=500)
@@ -185,8 +183,7 @@ def test_start_ibofos_already_running_failure(mock_get_current_user, mock_get_ib
 @mock.patch("rest.db.connection_factory.get_current_user",
             return_value="test", autospec=True)
 @mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
-@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
-def test_start_ibofos_failure(mock_get_current_user, mock_get_ibof_os_status,mock_is_ibofos_running, **kwargs):
+def test_start_ibofos_failure(mock_get_current_user, mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
                        json={"result": {"status": {"description": "FAIL", "code":20}}},
@@ -207,10 +204,9 @@ def test_start_ibofos_failure(mock_get_current_user, mock_get_ibof_os_status,moc
 
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
 @mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
 @mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
-def test_start_ibofos_already_running(mock_get_current_user,mock_get_ibof_os_statusi, mock_is_ibofos_running,  **kwargs):
+def test_start_ibofos_already_running(mock_get_current_user, mock_is_ibofos_running,  **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
                        json={"result": {"status": {"description": "SUCCESS", "code":0}}},
