+++ src/mtool/api/rest/blueprints/pos.py
@@ -1,63 +1,47 @@
-import os
 import datetime
+import json
+import re
+import os
+import uuid
+import rest.rest_api.dagent.ibofos as dagent
 from flask import Blueprint, make_response, request, jsonify
 from rest.auth import token_required
-from util.com.common import toJson
-from rest.log import logger
-import rest.rest_api.dagent.ibofos as dagent
-import json
-from rest.rest_api.system.system import fetch_system_state
 from rest.db import connection_factory
+from rest.log import logger
+from rest.rest_api.system.system import fetch_system_state
+from util.com.common import get_ip_address, get_hostname, toJson
 
 pos_bp = Blueprint('pos', __name__)
 
 class IBOF_OS_Running:
     Is_Ibof_Os_Running_Flag = False
 
-@pos_bp.route('/api/v1.0/version', methods=['GET'])
-def get_version():
-    package_json_path = os.path.abspath(
-        os.path.join(
-            os.path.dirname(__file__),
-            "../../../ui/package.json"))
-    with open(package_json_path, encoding="utf-8") as f:
-        data = json.load(f)
-        return toJson({"version": data["version"]})
+# Get IP And Mac Info
+@pos_bp.route('/api/v1.0/get_ip_and_mac', methods=['GET'])
+def getIpAndMac():
+    """
+    fetches ip and mac of the system
+    :return:
+    """
+    try:
+        ip = get_ip_address()
+        mac = ':'.join(re.findall('..', '%012x' % uuid.getnode()))
+        host = get_hostname()
+        if ip is None and mac == '':
+            return make_response('Could not get ip and mac', 500)
+        else:
+            lastUpdatedTime = datetime.datetime.now().strftime("%a, %d %b %Y %I:%M:%S %p %Z")
+            return jsonify({"ip": ip, "mac": mac, "host": host, "timestamp": lastUpdatedTime})
+    except BaseException:
+        return make_response('Could not get ip and mac', 500)
 
-
-@pos_bp.route('/api/v1.0/logger', methods=['POST'])
-def log_collect():
-    body_unicode = request.data.decode('utf-8')
-    if len(body_unicode) > 0:
-        body = json.loads(body_unicode)
-        try:
-            logger.info(body)
-        except Exception as e:
-            print(e)
-        return "Log written sucessfully"
-    return make_response("Received null log value", 200)
-
-
-def get_ibof_os_status():
-    return IBOF_OS_Running.Is_Ibof_Os_Running_Flag
-
+# Get POS Info
 @pos_bp.route('/api/v1.0/pos/info', methods=['GET'])
 @token_required
 def get_pos_info(current_user):
     return toJson(dagent.get_pos_info())
 
-@pos_bp.route('/api/v1.0/start_ibofos', methods=['GET'])
-@token_required
-def start_ibofos(current_user):
-    """
-    Start IBOF os pos_bplication. The path is given for demo&kntf server
-    :param current_user:
-    :return: status
-    """
-    res = dagent.start_ibofos()
-    res = res.json()
-    return jsonify(res)
-
+# Check POS Running
 @pos_bp.route('/api/v1.0/get_Is_Ibof_OS_Running/')
 @token_required
 def is_ibofos_running(current_user="admin"):
@@ -110,7 +94,20 @@ def is_ibofos_running(current_user="admin"):
                    "situation": situation,
                    "value": value})
 
+# Start POS
+@pos_bp.route('/api/v1.0/start_ibofos', methods=['GET'])
+@token_required
+def start_ibofos(current_user):
+    """
+    Start IBOF os pos_bplication. The path is given for demo&kntf server
+    :param current_user:
+    :return: status
+    """
+    res = dagent.start_ibofos()
+    res = res.json()
+    return jsonify(res)
 
+# Stop POS
 @pos_bp.route('/api/v1.0/stop_ibofos', methods=['GET'])
 @token_required
 def stop_ibofos(current_user):
